feat: dds ID validation to the assessment controller

+++ src/main/java/com/eng/homecare/controllers/AssessmentController.java
@@ -5,6 +5,7 @@ import com.eng.homecare.request.AssessmentRequestDTO;
 import com.eng.homecare.response.AssessmentResponseDTO;
 import com.eng.homecare.services.AssessmentService;
 import lombok.RequiredArgsConstructor;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
@@ -24,27 +25,28 @@ public class AssessmentController {
     }
 
     @GetMapping("{assessmentId}")
-    public ResponseEntity<AssessmentResponseDTO>listbyId(@PathVariable Long professionalId) {
-        return ResponseEntity.ok(assessmentService.listAssessment(professionalId));
+    public ResponseEntity<AssessmentResponseDTO>listbyId(@PathVariable Long assessmentId) {
+        return ResponseEntity.ok(assessmentService.listAssessment(assessmentId));
     }
 
-    @PatchMapping("/assessments/{id}")
+    @PatchMapping("/assessments/{assessmentId}")
     @PreAuthorize("hasRole('PATIENT')")
-    public ResponseEntity<AssessmentResponseDTO> patchAssessment(
-            @PathVariable Long id,
-            @RequestBody AssessmentRequestDTO dto) {
-
-        AssessmentResponseDTO response = assessmentService.patchAssessment(id, dto);
+    public ResponseEntity<AssessmentResponseDTO> patchAssessment(@PathVariable Long assessmentId,@RequestBody AssessmentRequestDTO dto, @AuthenticationPrincipal JWTUserData patientData) {
+        AssessmentResponseDTO existingAssessment = assessmentService.listAssessment(assessmentId);
+        if (!patientData.id().equals(existingAssessment.patientId())){
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+        }
+        AssessmentResponseDTO response = assessmentService.patchAssessment(assessmentId, dto);
         return ResponseEntity.ok(response);
     }
 
     @DeleteMapping("{assessmentId}")
     @PreAuthorize("hasRole('PATIENT')")
-    public ResponseEntity<String> deleteById(@PathVariable Long assessmentId) {
-//        JWTUserData userData = (JWTUserData) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
-//        if (!userData.id().equals(patientId)){
-//            return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
-//        }
+    public ResponseEntity<String> deleteById(@PathVariable Long assessmentId, @AuthenticationPrincipal JWTUserData patientData) {
+        AssessmentResponseDTO existingAssessment = assessmentService.listAssessment(assessmentId);
+        if (!patientData.id().equals(existingAssessment.patientId())){
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+        }
         assessmentService.deleteAssessment(assessmentId);
         return ResponseEntity.ok("Assessment deleted");
     }
